(After Merge) Fix sticky footer +add margin-bottom
Sticky footer now perfectly works. The error was due to an oversight of
an absolute property in #footer div.

Add of a margin bottom tom make more space between last post and footer

TO DO : Problem with add page : the page is not contain in hierachical
divs (wrap, contains) so does not take they properties